Moved new terms type files into `server/shared/types/terms`
