+++ server/shared/types/index.ts
@@ -4,8 +4,8 @@ export * from './termsetting'
 export * from './termdb'
 export * from './filter'
 export * from './vocab'
-export * from './categorical'
-export * from './condition'
-export * from './numeric'
-export * from './geneVariant'
-export * from './snps'
+export * from './terms/categorical'
+export * from './terms/condition'
+export * from './terms/numeric'
+export * from './terms/geneVariant'
+export * from './terms/snps'
