+++ server/shared/types/terms/snps.ts
@@ -1,7 +1,7 @@
-import { TermWrapper, BaseQ, Term } from './termdb'
-import { TermSettingInstance, InstanceDom, UseCase } from './termsetting'
-import { VocabApi } from './vocab'
-import { Tvs, Filter } from './filter'
+import { TermWrapper, BaseQ, Term } from '../termdb'
+import { TermSettingInstance, InstanceDom, UseCase } from '../termsetting'
+import { VocabApi } from '../vocab'
+import { Tvs, Filter } from '../filter'
 
 /*
 --------EXPORTED--------
